Major improvement

+++ src/devices-interfaces/low-level/PwmDevice.cpp
@@ -12,12 +12,12 @@ int8_t PwmDevice::init() {
     return 0;
 }
 
-int8_t PwmDevice::initPin(uint16_t channel, uint16_t pinNum, float pwmFreq) {
+int8_t PwmDevice::initPin(uint8_t channel, uint8_t pinNum, float pwmFreq) {
     ledcSetup(channel, pwmFreq, 12);
     ledcAttachPin(pinNum, channel);
     return 0;
 }
 
-void PwmDevice::setPin(uint16_t channel, uint16_t pinNum, uint16_t dutyCycle) {
+void PwmDevice::setPin(uint8_t channel, uint8_t pinNum, uint16_t dutyCycle) {
     ledcWrite(channel, dutyCycle);
 }
